Notebook - Do not duplicate the file on disk if a list item is cloned. Multiple references can point to the same storedFileId

+++ apphub-service/apphub-service-notebook/src/main/java/com/github/saphyra/apphub/service/notebook/service/FileDeletionService.java
@@ -17,7 +17,10 @@ public class FileDeletionService {
 
     public void deleteFile(UUID parent) {
         File file = fileDao.findByParentValidated(parent);
-        storageProxy.deleteFile(file.getStoredFileId());
+        if (fileDao.countByStoredFileId(file.getStoredFileId()) < 2) {
+            storageProxy.deleteFile(file.getStoredFileId());
+        }
+
         fileDao.delete(file);
     }
 }
